+++ ichnaea/api/locate/tests/test_cell.py
@@ -1,4 +1,5 @@
 from ichnaea.api.locate.cell import CellPositionSource
+from ichnaea.api.locate.result import ResultList
 from ichnaea.api.locate.tests.base import BaseSourceTest
 from ichnaea.constants import LAC_MIN_ACCURACY
 from ichnaea.tests.factories import (
@@ -14,7 +15,7 @@ class TestCellPosition(BaseSourceTest):
     def test_check_empty(self):
         query = self.model_query()
         result = self.source.result_type()
-        self.assertFalse(self.source.should_search(query, result))
+        self.assertFalse(self.source.should_search(query, ResultList(result)))
 
     def test_empty(self):
         query = self.model_query()
@@ -63,7 +64,8 @@ class TestCellPosition(BaseSourceTest):
         with self.db_call_checker() as check_db_calls:
             query = self.model_query(cells=cells)
             result = self.source.result_type()
-            self.assertFalse(self.source.should_search(query, result))
+            self.assertFalse(
+                self.source.should_search(query, ResultList(result)))
             check_db_calls(rw=0, ro=0)
 
     def test_smallest_area(self):