+++ ichnaea/api/locate/tests/test_ocid.py
@@ -1,4 +1,5 @@
 from ichnaea.api.locate.ocid import OCIDPositionSource
+from ichnaea.api.locate.result import ResultList
 from ichnaea.api.locate.tests.base import BaseSourceTest
 from ichnaea.tests.factories import (
     OCIDCellAreaFactory,
@@ -13,7 +14,7 @@ class TestOCIDPositionSource(BaseSourceTest):
     def test_check_empty(self):
         query = self.model_query()
         result = self.source.result_type()
-        self.assertFalse(self.source.should_search(query, result))
+        self.assertFalse(self.source.should_search(query, ResultList(result)))
 
     def test_empty(self):
         query = self.model_query()